Fix `TabletMode` auto configuration
- Remedy MRDGH2821/kde-auto-touch-mode#1 . Events
of the same device action in a events batch
(defined as <= 0.1 second between such events,
from empirical testing) only applies configuration
once.
- Filter out device events where the action isn't
`"added"` or `"removed"`
- Fix setting `TabletMode` configuration value.
The code of `toggle_mode()` function did not match
its description; device action was not considered
at all, instead `TabletMode` was always toggled.
If the configuration was modified externally (for
example via System Settings), or if the function
was invoked even number of times per device
addition or removal, the configuration state the
script assumed would be out of sync with the
actual state.
- Stop catching `KeyboardInterrupt`. This was done
to stop `MonitorObserver` background thread, which
is not used here anymore. Allow `KeyboardInterrupt`
to end the process.