Don't change working directory
Previously during execution, working directory was
changed into the directory of
`auto_toggle_touch_mode.py` script to allow it to
run `toggle_touch_mode.py` in the same directory
in a subprocess.
`toggle_touch_mode` is now imported as a module by
`auto_toggle_touch_mode.py` script. When running
as a script, Python automatically adds the script
directory (not the working directory!) to the
module search paths, making it unnecessary to
change working directory.